Kanban AI auto-doing, single-task drag reorder, multi-space chips, global Ctrl+Enter save
Four features:

- Auto-select doing: a magic button in the Doing column header opens a
  "what do you want to do?" modal; POST /api/tasks/auto-doing asks the AI
  (new select_tasks_with_ai seam reusing the raw-completion cleanify path,
  new task_selection.md prompt) which TODO tasks match the stated intent
  and moves them to doing (audited actor='ai'). A space-filtered board
  restricts the candidates to the selected spaces.

- Reorder overhaul: POST /api/tasks/reorder now nudges ONLY the dragged
  task ({task_id, priority}) instead of rewriting every task's priority
  from list order. The client slots the dragged card between its new
  neighbours with a fractional priority (Task.priority is now Float;
  SQLite type affinity keeps the existing INTEGER-declared prod column
  storing REALs, so no migration is needed). Kanban columns now support
  intra-column drag reorder (Done stays completion-time ordered);
  cross-column drags keep their status-only semantics untouched. The
  calendar sidebar list uses the same single-task nudge. UI keeps showing
  rounded integer priorities.

- Multi-space viewing: Ctrl+click on a space chip toggles the space in/out
  of the filter (plain click still selects one, All resets) — on the
  kanban board and now also in Notes, where the same chips replace the old
  dropdown. GET /api/notes accepts repeated space_id params; note rows
  show a space tag when several spaces are visible; new notes land in the
  first selected space.

- Global Ctrl+Enter save: a capture-phase handler saves from wherever
  you're typing — clicks the open modal's primary action, flushes the
  notes autosave immediately (NotesView.saveNow), saves the space editor;
  capture inputs keep their Enter-creates behavior. Modal hides after
  fast submits are deferred past Bootstrap's show transition
  (hideModalSafely) so they can't be silently dropped.

Help modal and PROJECT_DESCRIPTION.md updated; 17 new route/seam tests
(91 total passing); verified end-to-end in a headless browser.

+++ src/models.py
@@ -19,7 +19,12 @@ class Task(db.Model):
     # is additive-only; its data is backfilled into space_id by a data fixup.
     space = db.Column(db.String(100))
     space_id = db.Column(db.Integer, db.ForeignKey('spaces.id'))  # Reference to Space table
-    priority = db.Column(db.Integer, default=0)  # Higher number = higher priority
+    # Higher number = higher priority. Float: user-facing editing stays on the
+    # integer 0-10 scale, but manual drag-reorder nudges the dragged task to a
+    # fractional value between its new neighbours so ONLY that task changes.
+    # (SQLite type affinity keeps prod's INTEGER-declared column storing
+    # fractional REALs losslessly — no migration needed.)
+    priority = db.Column(db.Float, default=0)
     deadline = db.Column(db.DateTime)
     estimated_duration = db.Column(db.Integer)  # in minutes
     scheduled_start = db.Column(db.DateTime)
